+++ scilab/modules/xml/sci_gateway/cpp/sci_xmlValidate.cpp
@@ -139,7 +139,7 @@ int sci_xmlValidate(char * fname, unsigned long fname_len)
 
         if (clines.size())
         {
-            err = createMatrixOfString(pvApiCtx, Rhs + 1, lines.size(), 1, const_cast<const char * const *>(&(clines[0])));
+            err = createMatrixOfString(pvApiCtx, Rhs + 1, (int)lines.size(), 1, const_cast<const char * const *>(&(clines[0])));
         }
         else
         {
